Fixed compile warning in tools:
/data/action/runner1/_work/ncnn/ncnn/tools/modelwriter.h:2212:35: warning: format ‘%llu’ expects argument of type ‘long long unsigned int’, but argument 3 has type ‘uint64_t’ {aka ‘long unsigned int’} [-Wformat=]
      |         fprintf(stderr, "mac = %llu = %.2f M\n", mac, mac / 1000000.0);
      |                                ~~~^              ~~~
      |                                   |              |
      |                                   |              uint64_t {aka long unsigned int}
      |                                   long long unsigned int
      |                                %lu

+++ tools/modelwriter.h
@@ -2209,7 +2209,7 @@ int ModelWriter::save(const char* parampath, const char* binpath)
 
     if (mac)
     {
-        fprintf(stderr, "mac = %llu = %.2f M\n", mac, mac / 1000000.0);
+        fprintf(stderr, "mac = %llu = %.2f M\n", static_cast<long long unsigned>(mac), mac / 1000000.0);
     }
 
     return 0;
